[level 3] Title: 광고 삽입, Time: 17.16 ms, Memory: 80.8 MB -BaekjoonHub

+++ 프로그래머스/3/72414. 광고 삽입/광고 삽입.java
@@ -32,7 +32,7 @@ class Solution {
             interval[e]--;
         }
         
-        int val = interval[0];
+        long val = interval[0];
         for (int i = 1; i <= pt; i++) {
             prefixed[i] += val + prefixed[i - 1];
             val += interval[i];
